C++ Code Sample of an S parameter Simulation
This code features a demonstration of a class used to simulate two-port
RF circuits. With this class, a user can build circuits from passive
components and simulate the S parameters in order to characterize the
frequency response of the circuit. Furthermore, the center  requency and
insertion loss of a filter design may be calculated.

Included here is the class as well as a demonstration of a simulation of
a third-order filter composed of two-port contour-mode resonators. The
user will prompted to give a quality factor value, which will be used in
the simulation of the filter. A typical quality factor is ~1500. The
user will be given the option of saving the simulated S parameters of
|S21| values in tab-delineated files, which can easily be opened in
excel to visualize the frequency response of the circuit.